requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Financial data
yfinance>=0.1.87
pandas-datareader>=0.10.0

# Machine Learning
xgboost>=1.6.0
lightgbm>=3.3.0

# Statistical analysis
statsmodels>=0.13.0
arch>=5.3.0

# Web scraping and APIs
requests>=2.28.0
beautifulsoup4>=4.11.0
scrapy>=2.6.0

# Natural Language Processing
nltk>=3.7.0
textblob>=0.17.0
vaderSentiment>=3.3.0

# Configuration and utilities
python-dotenv>=0.19.0
pyyaml>=6.0
jupyter>=1.0.0
jupyterlab>=3.4.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
